-> Worked on the task detail view. -> Users can easily mark the status of a task -> Worked on deleting modal a bit to use it with the tasks detail view as well (incomplete, will complete in next session)

+++ frontend/src/Components/DeleteModal.js
@@ -2,21 +2,26 @@ import React from "react";
 import { useProjectContext } from "../Context/ProjectContext";
 import axios from "axios";
 
-export default function DeleteModal({ setShow, navigate, id })  {
+export default function DeleteModal({ setShow, navigate, id, task })  {
     const {token} = useProjectContext();
-    const hanldeDeleteProject = async() => {
-        await axios.delete(`http://127.0.0.1:8000/api/projects/${id}/`, {
-            headers: {
-                Authorization: `Token ${token}`
-            }
-        })
-        .then((response) => {
-            console.log(response.data);
-            navigate('/');
-        })
-        .catch((error) => {
-            console.error(error);
-        })
+    const handleDelete = async() => {
+        if (task) {
+
+        }
+        else {
+            await axios.delete(`http://127.0.0.1:8000/api/projects/${id}/`, {
+                headers: {
+                    Authorization: `Token ${token}`
+                }
+            })
+            .then((response) => {
+                console.log(response.data);
+                navigate('/');
+            })
+            .catch((error) => {
+                console.error(error);
+            })
+        }
     }
 
     const hanldeClickOutside = (e) => {
@@ -29,9 +34,9 @@ export default function DeleteModal({ setShow, navigate, id })  {
         <div className="modal-container-delete" onClick={hanldeClickOutside}>
             <div className="modal-content-delete">
                 <h2>Confirm Delete</h2>
-                <p>Are you sure you want to delete this project?</p>
+                <p>Are you sure you want to delete this {task?'task' : 'project'}?</p>
                 <div className="delete-btn-container">
-                    <button className="btn btn-confirm" onClick={hanldeDeleteProject}>
+                    <button className="btn btn-confirm" onClick={handleDelete}>
                         Confirm
                     </button>
                     <button className="btn btn-cancel" onClick={() => setShow(false)}>
@@ -41,4 +46,8 @@ export default function DeleteModal({ setShow, navigate, id })  {
             </div>
         </div>
     )
+}
+
+DeleteModal.defaultProps = {
+    task: false,
 }
